requirements.txt
Flask==2.3.3
flask-cors==3.0.10
pdfminer.six==20221105
google-generativeai==0.2.0